==> notebooks/03_conversation_executive_wise.ipynb <==
import os 
import sys
import json 
import pandas as pd 
import re

from langchain.document_loaders import JSONLoader,TextLoader
from langchain.text_splitter import CharacterTextSplitter

os.environ['OPENAI_API_KEY'] = ""

### Data Mortgage 

excel_path = "..//data/Data_Mortgage.xlsx"
excel_data = pd.read_excel(excel_path)
excel_data.dropna(subset='Opportunity',inplace=True)
excel_data.head()

### Read The Json data of recordings
data_path = "../data/Processed_data/Audio_data.json"
with open(data_path,'r+') as fd:
    audio_data = json.load(fd)
fd.close()

### Extract all the records for Sales Executive

For example : ABC had converastions with multiple customers so giving this record to data Loader 

name_of_executive = 'Juraira Manzoor'
df_executive = excel_data[excel_data['Assigned']==name_of_executive]
print(df_executive.shape)
df_executive.head()

conversation_dict = dict()
conversation_dict['data']=list()
Conversation = []
customers = []

for records in audio_data['data']:
    if records['relationship_manager'] == 'Jurara':
        text = re.sub("\s\s+", " ", records['text'])
        customer = records['customer']
        text = "conversation with Customer {}:{}".format(customer,text)

        conversation_dict['data'].append({
                'name':name_of_executive,
                'text':text,
                'customer':records['customer'],
                'language':records['language']
            })

conversation_dict

with open('../data/tmp.json','w+') as fd:
    json.dump(conversation_dict,fd)


### Load data as JsonLoader



def metadata_func(record: dict, metadata: dict) -> dict:
    metadata['customer'] = record.get('customer')
    metadata['language'] = record.get('language')
    return metadata

loader = JSONLoader(
    file_path='../data/tmp.json',
    jq_schema='.data[]',
    content_key="text",
    metadata_func=metadata_func
)
texts = loader.load()


### Load Embeddings to Vector Database

import openai
import os 
from langchain.embeddings.base import Embeddings
from typing import List
from sentence_transformers import SentenceTransformer
from langchain.embeddings.openai import OpenAIEmbeddings

class LocalHuggingFaceEmbeddings(Embeddings):
    def __init__(self, model_id): 
        # Should use the GPU by default
        self.model = SentenceTransformer(model_id)
        
    def embed_documents(self, texts: List[str]) -> List[List[float]]:
        """Embed a list of documents using a locally running
           Hugging Face Sentence Transformer model
        Args:
            texts: The list of texts to embed.
        Returns:
            List of embeddings, one for each text.
        """
        embeddings =self.model.encode(texts)
        return embeddings

    def embed_query(self, text: str) -> List[float]:
        """Embed a query using a locally running HF 
        Sentence trnsformer. 
        Args:
            text: The text to embed.
        Returns:
            Embeddings for the text.
        """
        embedding = self.model.encode(text)
        return list(map(float, embedding))

huggingface_embeddings = LocalHuggingFaceEmbeddings('multi-qa-mpnet-base-dot-v1')

from langchain.vectorstores import FAISS

vectorstore =  FAISS.from_documents(documents=texts,
                                    embedding = huggingface_embeddings
                                    ) # turn dcos into Vectors and store them in RAM also add metadata 
vectorstore.save_local('../data/faiss_dmac_gpt_exp')

### Chat Query with ConversationalRetrievalChain

import os
from dotenv import load_dotenv
from langchain.llms import OpenAI
from langchain.prompts import PromptTemplate
from langchain.chains.question_answering import load_qa_chain
from langchain.chains import ConversationalRetrievalChain
from langchain.vectorstores import FAISS
from langchain.chains import LLMChain
from langchain.chains.conversational_retrieval.prompts import CONDENSE_QUESTION_PROMPT

chat = OpenAI(temperature=0)
# memory = ConversationBufferMemory(memory_key="chat_history", return_messages=True)
vectordb = FAISS.load_local('../data/faiss_dmac_gpt_exp/',embeddings=huggingface_embeddings)

### Create Prompt

def create_prompt():
    prompt_template = """
    Analyze conversations between customer and sales executive from context.
    If customer shows interest in service or Property , conversation is a potential lead.  
    Always answer point wise with person names. Don't make up answers
   
    {context}
   
    {chat_history}
   
    Question: {question}
    Answer stepwise: 
    """
    prompt = PromptTemplate(input_variables=["context", "question", "chat_history"], template=prompt_template)
    return prompt


question_generator = LLMChain(llm=chat, prompt=create_prompt())
doc_chain = load_qa_chain(chat,chain_type='map_reduce')

chain = ConversationalRetrievalChain.from_llm(
                                    llm = chat,
                                    retriever=vectordb.as_retriever(),
                                    # question_generator=question_generator,
                                    combine_docs_chain_kwargs={"prompt": create_prompt()},
                                    )
chat_history = []


### Query with customers 

query = "What is Ms. Emmy Interested in? "
result = chain({"question": query, "chat_history": chat_history})

print("query:{}".format(result['question']))
print("response:\n{}".format(result['answer']))
chat_history.append((query,result['answer']))

print(chat_history)

query = "How much mortage is she talking about ?"
result = chain({"question": query, "chat_history": chat_history})
print("query:{}".format(result['question']))
print("response:\n{}".format(result['answer']))
chat_history.append((query,result['answer']))

query = "What are the next steps for her  ?"
result = chain({"question": query, "chat_history": chat_history})
print("query:{}".format(result['question']))
print("response:\n{}".format(result['answer']))
chat_history.append((query,result['answer']))

query = "What is the value he looking for ?"
result = chain({"question": query, "chat_history": chat_history})
print("query:{}".format(result['question']))
print("response:\n{}".format(result['answer']))
chat_history.append((query,result['answer']))

from langchain.chat_models import ChatOpenAI
from langchain.embeddings.openai import OpenAIEmbeddings
docsearch = FAISS.load_local("../vector_embeddings/faiss_damac_audio_data_openai/", embeddings=OpenAIEmbeddings())
prompt = create_prompt()
chat = ChatOpenAI(verbose=True, temperature=0)

qa = ConversationalRetrievalChain.from_llm(llm=chat,
                                            retriever=docsearch.as_retriever(),
                                            combine_docs_chain_kwargs={"prompt": prompt},
                                            max_tokens_limit=4097
                                            )

chat_history = []
generated_response = qa({'question':'summary of conversations',"chat_history":chat_history})
formatted_response = f"{generated_response['answer']} \n\n"


print(formatted_response)

from langchain.chat_models import ChatOpenAI
from langchain.embeddings.openai import OpenAIEmbeddings
docsearch = FAISS.load_local("../vector_embeddings/faiss_damac_audio_data_hugging_face/", 
                             embeddings=huggingface_embeddings)
prompt = create_prompt()
chat = ChatOpenAI(verbose=True, temperature=0)

qa = ConversationalRetrievalChain.from_llm(llm=chat,
                                            retriever=docsearch.as_retriever(),
                                            combine_docs_chain_kwargs={"prompt": prompt},
                                            max_tokens_limit=4097
                                            )

chat_history = []
generated_response = qa({'question':'summary of conversations',"chat_history":chat_history})
formatted_response = f"{generated_response['answer']} \n\n"
print(formatted_response)

# from glob import glob 

# filenames = glob("../data/Audio/*.json")


# def merge_JsonFiles(filename):
#     result = {'conversation data':[]}
#     for f1 in filename:
#         with open(f1, 'r') as infile:
#             data = json.load(infile)
#         print(f1,len(data['data']))
#         # result['conversation data'].append(data['data'][0])

#     # with open('counseling3.json', 'w') as output_file:
#     #     json.dump(result, output_file,indent=4)

# merge_JsonFiles(filenames)
